button para deleta o card register do cliente e no banco authentication

+++ src/app/pages/card-id/card-id.component.ts
@@ -2,6 +2,8 @@ import { FormGroup } from '@angular/forms';
 import { Component, EventEmitter } from '@angular/core';
 import { CardService } from 'src/app/services/card.service';
 import { Register } from 'src/app/interface/register.interface';
+import Swal from 'sweetalert2';
+import { AuthService } from 'src/app/auth/auth.service';
 
 @Component({
   selector: 'app-card-id',
@@ -13,7 +15,7 @@ export class CardIDComponent {
 
   cardObjs: Register[] = []
 
-  constructor(private cardService: CardService) {
+  constructor(private cardService: CardService, private afAuth: AuthService) {
     this.cardService.getData('').subscribe(data => {
       this.cardObjs = data;
       if (data.length > 0) {
@@ -24,12 +26,25 @@ export class CardIDComponent {
 
   editar() { }
 
-  deleteCard(uid: string): void {
-    this.cardService.deleteData(uid);
-    if (this.cardService) {
-      alert('Deletado com sucesso')
-    } else {
-      alert('Delete error');
-    }
+  deleteCard(uid: string, name: string): void {
+    Swal.fire({
+      title: `Você tem certeza quer deletar usuário ( ${name} ) do banco de dados!.`,
+      text: "Você não será capaz de reverter isso!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, deletar !'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.cardService.deleteData(uid);
+        this.afAuth.deleteDoc(uid);
+        Swal.fire(
+          'Deleted!',
+          `usuário deletado com sucesso ${name}`,
+          'success'
+        )
+      }
+    })
   }
 }
